+++ src/tree/zk.rs
@@ -5,8 +5,8 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 use super::PoseidonBranch;
+use dusk_hades::{GadgetStrategy, Strategy};
 use dusk_plonk::prelude::*;
-use hades252::{GadgetStrategy, Strategy};
 
 /// Perform a merkle opening for a given branch and return the calculated root
 pub fn merkle_opening<const DEPTH: usize>(
@@ -19,13 +19,13 @@ pub fn merkle_opening<const DEPTH: usize>(
     let mut root = zero;
 
     // Generate a permutation container
-    let mut perm = [zero; hades252::WIDTH];
+    let mut perm = [zero; dusk_hades::WIDTH];
 
     // For every level, replace the level offset with needle,
     // permutate the level and set the needle to the next level
     // to the poseidon result of the permutation
     branch.as_ref().iter().fold(leaf, |needle, level| {
-        let offset = level.offset();
+        let offset = level.offset() as usize;
 
         level.as_ref().iter().enumerate().for_each(|(i, l)| {
             if i != offset {